feat: refactored codebase as per common react code structure
