+++ babel.config.js
@@ -1,6 +1,8 @@
 module.exports = {
-    presets: ['@babel/preset-env', '@babel/preset-react',  '@babel/preset-typescript'],
-    "plugins": [
-        ["@babel/plugin-transform-runtime"]
-    ]
+  presets: [
+    "@babel/preset-env",
+    "@babel/preset-react",
+    "@babel/preset-typescript",
+  ],
+  plugins: [["@babel/plugin-transform-runtime"]],
 };
