+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-custom-popup",
-  "version": "1.2.1",
+  "version": "1.3.0",
   "homepage": "https://Rafaa17.github.io/react-custom-modal",
   "description": "React Popup Library build using React Context API",
   "main": "build/index.js",
@@ -19,11 +19,12 @@
     "url": "https://github.com/Rafaa17/react-custom-modal"
   },
   "scripts": {
-    "test": "jest --watchAll",
+    "test": "jest --watchAll -",
     "build-demo": "rm -rf build build-demo && react-scripts build && mv build build-demo",
     "deploy-demo": "gh-pages -d build-demo",
     "build": "rm -rf build && rollup -c",
-    "start": "HOST=0.0.0.0 PORT=3003 react-scripts start"
+    "start": "HOST=0.0.0.0 PORT=3003 craco start",
+    "typecheck": "tsc --noEmit"
   },
   "author": "",
   "license": "ISC",
@@ -44,7 +45,10 @@
     "@babel/core": "^7.11.6",
     "@babel/plugin-transform-runtime": "^7.17.0",
     "@babel/preset-typescript": "^7.16.7",
+    "@craco/craco": "^7.1.0",
+    "@craco/types": "^7.1.0",
     "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.5.2",
     "@types/jest": "^29.4.0",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
@@ -85,7 +89,8 @@
   },
   "jest": {
     "moduleNameMapper": {
-      "\\.(css|less|scss|sass)$": "identity-obj-proxy"
+      "\\.(css|less|scss|sass)$": "identity-obj-proxy",
+      "^@/(.*)$": "<rootDir>/src/$1"
     },
     "testEnvironment": "jsdom"
   }
