+++ src/hooks/usePopup.ts
@@ -1,0 +1,5 @@
+import { useContext } from "react";
+
+import { PopupContext } from "@/context/PopupContext";
+
+export const usePopup = () => useContext(PopupContext);
