+++ tsconfig.json
@@ -3,11 +3,7 @@
     "outDir": "build",
     "module": "esnext",
     "target": "es6",
-    "lib": [
-      "es5",
-      "dom",
-      "es2016"
-    ],
+    "lib": ["es5", "dom", "es2016"],
     "sourceMap": true,
     "allowJs": false,
     "jsx": "react",
@@ -19,7 +15,6 @@
     "noImplicitAny": true,
     "allowSyntheticDefaultImports": true,
     "strictNullChecks": true,
-    "suppressImplicitAnyIndexErrors": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "downlevelIteration": true,
@@ -28,13 +23,11 @@
     "strict": true,
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "noEmit": true
+    "noEmit": true,
+    "paths": {
+      "@/*": ["./src/*"]
+    }
   },
-  "include": [
-    "src"
-  ],
-  "exclude": [
-    "node_modules",
-    "build"
-  ]
+  "include": ["./src/", "./__tests__/"],
+  "exclude": ["node_modules", "build"]
 }
